refactor(session): extract turn history recovery
